Split out heart room to .h file on PSX
