+++ src/st/np3/e_heartroom.c
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "np3.h"
+#include "../heartroom_psx.h"
